+++ app/(main)/(tabs)/(goals)/goalsResumen.tsx
@@ -1,0 +1,2 @@
+export { default } from "@/modules/goals/screens/GoalsScreen";
+