+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2025-10-YY
+
+### Added
+
+- Goals Module
+
+### Chaged
+
+-stting screen and tab, for logout tab
+
+### Fixed
+
 ## [0.0.3](https://github.com/JavierTaborda/AppFrigilux/releases/tag/v0.0.3) - 2025-10-10
 
 ### Added